+++ app/routes/auth.routes.js
@@ -1,26 +1,25 @@
 const {
     Router,
 } = require('express');
+const passport = require('passport');
 
-const passport=require('passport');
 const init = (app, data) => {
     const router = new Router();
-    console.log('auth route!');
-    router.get('/login', (req, res)=>{
-    //    res.render('_shared/login');
-     res.redirect('/');
+    console.log('auth routes');
+    router.get('/login', (req, res) => {
+        res.redirect('/');
     });
-
-    router.post('/login',
+    router.post('/login', (req, res) => {
         passport.authenticate('local', {
             successRedirect: '/',
             failureRedirect: '/',
             failureFlash: true,
-        })
-    );
+        });
+    });
+    app.use('/', router);
 };
 
 
-module.exports={
+module.exports = {
     init,
 };
